# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Physical Venue Risk Assessment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues:
    get:
      operationId: listVenues
      tags: [venues]
      summary: List all venues
      responses:
        "200":
          description: List of venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
    post:
      operationId: createVenue
      tags: [venues]
      summary: Create a venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
  /venues/{id}:
    get:
      operationId: getVenue
      tags: [venues]
      summary: Get venue detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueDetail"
        "404":
          description: Not found
    patch:
      operationId: updateVenue
      tags: [venues]
      summary: Update a venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "404":
          description: Not found
    delete:
      operationId: deleteVenue
      tags: [venues]
      summary: Delete a venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Venue:
      type: object
      required: [id, name, venueType, address, city, country, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        venueType:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        environmentType:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        assessmentCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    VenueDetail:
      allOf:
        - $ref: "#/components/schemas/Venue"
        - type: object
          properties:
            assessments:
              type: array
              items:
                $ref: "#/components/schemas/AssessmentSummary"
            recentIncidents:
              type: array
              items:
                $ref: "#/components/schemas/Incident"
    VenueInput:
      type: object
      required: [name, address, city, country]
      properties:
        name:
          type: string
        venueType:
          type: string
          enum: [hotel, stadium, conference, embassy, airport, mall, government, restaurant, hospital, university, other]
        address:
          type: string
        city:
          type: string
        country:
          type: string
        lat:
          type: number
        lng:
          type: number
        googleMapsUrl:
          type: string
        district:
          type: string
        environmentType:
          type: string
          enum: [urban, suburban, rural, industrial, coastal]
        notes:
          type: string
    VenueUpdate:
      type: object
      properties:
        name:
          type: string
        venueType:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        lat:
          type: number
        lng:
          type: number
        googleMapsUrl:
          type: string
        district:
          type: string
        environmentType:
          type: string
        notes:
          type: string
    AssessmentSummary:
      type: object
      required: [id, title, status, version, createdAt, updatedAt]
      properties:
        id:
          type: integer
        venueId:
          type: ["integer", "null"]
        venueName:
          type: ["string", "null"]
        venueCity:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, under_review, approved, monitoring, review_required, escalated, archived]
        version:
          type: integer
        overallRating:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Incident:
      type: object
      required: [id, incidentType, severity, incidentDate, summary, verified, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: ["integer", "null"]
        venueName:
          type: ["string", "null"]
        incidentType:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        incidentDate:
          type: string
        summary:
          type: string
        sourceName:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        distanceFromVenue:
          type: ["number", "null"]
        confidenceLevel:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
...