# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Physical Venue Risk Assessment API
 * OpenAPI spec version: 0.2.0
 */

export interface Venue {
  id: number;
  name: string;
  venueType: string;
  address: string;
  city: string;
  country: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  /** @nullable */
  googleMapsUrl?: string | null;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  environmentType?: string | null;
  /** @nullable */
  notes?: string | null;
  assessmentCount?: number;
  createdAt: string;
  updatedAt: string;
}