# lib/api-zod/src/generated/types/venueUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Physical Venue Risk Assessment API
 * OpenAPI spec version: 0.2.0
 */

export interface VenueUpdate {
  name?: string;
  venueType?: string;
  address?: string;
  city?: string;
  country?: string;
  lat?: number;
  lng?: number;
  googleMapsUrl?: string;
  district?: string;
  environmentType?: string;
  notes?: string;
}